[monarch] Move hyperactor_mesh::v1::* into hyperactor_mesh::*
Eliminates the v1 versioning by moving all content from `hyperactor_mesh::v1`
into the base `hyperactor_mesh` module.

Changes:
- Moved files from src/v1/ to src/ (actor_mesh.rs, host_mesh.rs, proc_mesh.rs,
  mesh_controller.rs, namespace.rs, testactor.rs, testing.rs, value_mesh.rs)
- Merged v1.rs content into lib.rs (Error enum, CodecError, Name enum,
  NameParseError, Result type alias, StatusMesh/StatusOverlay type aliases)
- Deleted the v1 module (v1.rs and v1/ directory)
- Updated all internal imports from crate::v1:: to crate::
- Updated all external imports from hyperactor_mesh::v1:: to hyperactor_mesh::

Files updated across: hyperactor_mesh, monarch_hyperactor, monarch_extension,
hyper, monarch_rdma (examples, tests, benches)

Differential Revision: [D92343626](https://our.internmc.facebook.com/intern/diff/D92343626/)

[ghstack-poisoned]
