+++ hyperactor_mesh/examples/test_bench.rs
@@ -23,11 +23,11 @@ use hyperactor::PortRef;
 use hyperactor::Unbind;
 use hyperactor::clock::Clock;
 use hyperactor::clock::RealClock;
+use hyperactor_mesh::actor_mesh::ActorMesh;
 use hyperactor_mesh::bootstrap::BootstrapCommand;
 use hyperactor_mesh::comm::multicast::CastInfo;
-use hyperactor_mesh::proc_mesh::global_root_client;
-use hyperactor_mesh::v1::actor_mesh::ActorMesh;
-use hyperactor_mesh::v1::host_mesh::HostMesh;
+use hyperactor_mesh::global_root_client;
+use hyperactor_mesh::host_mesh::HostMesh;
 use ndslice::Point;
 use ndslice::ViewExt;
 use ndslice::extent;
