+++ hyperactor_mesh/src/alloc.rs
@@ -42,7 +42,7 @@ use typeuri::Named;
 
 use crate::alloc::test_utils::MockAllocWrapper;
 use crate::assign::Ranks;
-use crate::proc_mesh::mesh_agent::ProcMeshAgent;
+use crate::mesh_agent::ProcMeshAgent;
 use crate::shortuuid::ShortUuid;
 
 /// Errors that occur during allocation operations.
@@ -636,9 +636,9 @@ pub(crate) mod testing {
     use super::*;
     use crate::alloc::test_utils::TestActor;
     use crate::alloc::test_utils::Wait;
-    use crate::proc_mesh::default_transport;
-    use crate::proc_mesh::mesh_agent::GspawnResult;
-    use crate::proc_mesh::mesh_agent::MeshAgentMessageClient;
+    use crate::mesh_agent::GspawnResult;
+    use crate::mesh_agent::MeshAgentMessageClient;
+    use crate::transport::default_transport;
 
     #[macro_export]
     macro_rules! alloc_test_suite {
