+++ hyperactor_mesh/src/bootstrap/mailbox.rs
@@ -22,7 +22,7 @@ use hyperactor::mailbox::MailboxSender;
 use hyperactor::mailbox::MessageEnvelope;
 use hyperactor::mailbox::Undeliverable;
 
-use crate::v1;
+use crate::Name;
 
 /// LocalProcDialer dials local procs directly through a configured socket
 /// directory.
@@ -64,7 +64,7 @@ impl MailboxSender for LocalProcDialer {
             && addr == &self.local_addr
             // ...and only non-system procs on that address; the rest are directly
             // reachable through the backend address.
-            && name.parse::<v1::Name>().as_ref().is_ok_and(v1::Name::is_suffixed)
+            && name.parse::<Name>().as_ref().is_ok_and(Name::is_suffixed)
         {
             let senders = self.local_senders.read().unwrap();
             let senders = if senders.contains_key(name) {
@@ -117,7 +117,7 @@ mod tests {
     use hyperactor_config::attrs::Attrs;
 
     use super::*;
-    use crate::v1::Name;
+    use crate::Name;
 
     #[tokio::test]
     async fn test_proc_dialer() {
