+++ hyperactor_mesh/test/hyperactor_mesh_proxy_test.rs
@@ -22,13 +22,13 @@ use hyperactor::PortRef;
 use hyperactor::RemoteSpawn;
 use hyperactor::channel::ChannelTransport;
 use hyperactor::context::Mailbox;
+use hyperactor_mesh::ActorMesh;
+use hyperactor_mesh::ProcMesh;
 use hyperactor_mesh::alloc::AllocSpec;
 use hyperactor_mesh::alloc::Allocator;
 use hyperactor_mesh::alloc::ProcessAllocator;
-use hyperactor_mesh::proc_mesh::global_root_client;
+use hyperactor_mesh::global_root_client;
 use hyperactor_mesh::supervision::MeshFailure;
-use hyperactor_mesh::v1::ActorMesh;
-use hyperactor_mesh::v1::ProcMesh;
 use ndslice::View;
 use ndslice::extent;
 use serde::Deserialize;
