+++ monarch_extension/src/mesh_controller.rs
@@ -37,9 +37,9 @@ use hyperactor::actor::ActorStatus;
 use hyperactor::context;
 use hyperactor::mailbox::MailboxSenderError;
 use hyperactor::supervision::ActorSupervisionEvent;
+use hyperactor_mesh::ActorMesh;
+use hyperactor_mesh::ProcMeshRef;
 use hyperactor_mesh::supervision::MeshFailure;
-use hyperactor_mesh::v1::ActorMesh;
-use hyperactor_mesh::v1::ProcMeshRef;
 use monarch_hyperactor::actor::PythonMessage;
 use monarch_hyperactor::actor::PythonMessageKind;
 use monarch_hyperactor::buffers::Buffer;
